+++ src/client/src/containers/AppServiceModal/AppServicePlanInfo/AppServicePlanInfo.test.tsx
@@ -5,7 +5,7 @@ import { getInitialState, setSubscriptions } from "../../../mockData/mockStore";
 import { RenderResult } from "@testing-library/react";
 import { renderWithStore } from "../../../testUtils";
 import messages from "./messages";
-import { AZURE_LINKS } from "../../../utils/constants";
+import { AZURE_LINKS } from "../../../utils/constants/constants";
 
 describe("AppServicePlanInfo", () => {
   let props: any;
