+++ src/client/src/containers/AzureLoginModal/index.tsx
@@ -13,10 +13,10 @@ import classnames from "classnames";
 import { FormattedMessage } from "react-intl";
 import keyUpHandler from "../../utils/keyUpHandler";
 import messages from "./messages";
-import { KEY_EVENTS } from "../../utils/constants";
+import { KEY_EVENTS } from "../../utils/constants/constants";
 import { ReactComponent as Cancel } from "../../assets/cancel.svg";
 import CollapsibleInfoBox from "../../components/CollapsibleInfoBox";
-import { WIZARD_CONTENT_INTERNAL_NAMES } from "../../utils/constants";
+import { WIZARD_CONTENT_INTERNAL_NAMES } from "../../utils/constants/constants";
 import * as ModalActions from "../../store/navigation/modals/action";
 import { azureLogin } from "../../utils/extensionService/extensionService";
 import { AppContext } from "../../AppContext";
