+++ src/client/src/containers/GenerationModal/GenerationItem/index.tsx
@@ -17,7 +17,7 @@ import {
 } from "../../../utils/extensionService/extensionService";
 import { AppContext } from "../../../AppContext";
 import { useEffect } from "react";
-import { EXTENSION_COMMANDS } from "../../../utils/constants";
+import { EXTENSION_COMMANDS } from "../../../utils/constants/constants";
 
 interface IProps {
   item: GenerationItemData;
