+++ src/client/src/containers/PageSelectFrameworks/FrameworkCard/DependencyInfo/index.tsx
@@ -4,7 +4,7 @@ import classnames from "classnames";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import {
   WIZARD_CONTENT_INTERNAL_NAMES
-} from "../../../../utils/constants";
+} from "../../../../utils/constants/constants";
 import Notification from "../../../../components/Notification";
 import messages from "./messages";
 import { DependencyContext } from "../../DependencyContext";
