+++ src/client/src/containers/PageSelectFrameworks/FrameworkCard/index.tsx
@@ -7,7 +7,8 @@ import styles from "./styles.module.css";
 import { getSvg } from "../../../utils/getSvgUrl";
 import DependencyInfo from "./DependencyInfo";
 import messages from "./messages";
-import { KEY_EVENTS, ROUTES } from "../../../utils/constants";
+import { KEY_EVENTS } from "../../../utils/constants/constants";
+import { ROUTES } from "../../../utils/constants/routes";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import { ReactComponent as Check } from "../../../assets/check.svg";
 import { getLatestVersion } from "../../../utils/extensionService/extensionService";
