+++ src/client/src/containers/RightSidebar/ServicesList/SidebarItem/index.tsx
@@ -5,7 +5,7 @@ import loadable from '@loadable/component'
 import { ReactComponent as CloseSVG } from "../../../../assets/cancel.svg";
 import { ISelected } from "../../../../types/selected";
 import styles from "./styles.module.css";
-import { KEY_EVENTS } from "../../../../utils/constants";
+import { KEY_EVENTS } from "../../../../utils/constants/constants";
 import { injectIntl, InjectedIntl, InjectedIntlProps } from "react-intl";
 import { AppState } from "../../../../store/combineReducers";
 import messages from "./messages";
