+++ src/client/src/store/templates/features/selector.ts
@@ -3,7 +3,7 @@ import { AppState } from "../../combineReducers";
 import { IOption } from "../../../types/option";
 import { openAppServiceModalAction, openCosmosDbModalAction } from "../../navigation/modals/action";
 import messages from "./messages";
-import { SERVICE_GROUPS, WIZARD_CONTENT_INTERNAL_NAMES } from "../../../utils/constants";
+import { SERVICE_GROUPS, WIZARD_CONTENT_INTERNAL_NAMES } from "../../../utils/constants/constants";
 
 const getFeatures = (state: AppState) => state.templates.featureOptions;
