+++ src/client/src/utils/extensionService/extensionService.ts
@@ -1,7 +1,7 @@
 import { IVSCodeObject } from "../../types/vscode";
 import {
   EXTENSION_COMMANDS, EXTENSION_MODULES, WIZARD_CONTENT_INTERNAL_NAMES, PAYLOAD_MESSAGES_TEXT
-} from "../constants";
+} from "../constants/constants";
 import { ILoggingPayload } from "../../types/logger";
 
 const postMessageAsync = (command: string, paramsMessage: any, vscode: IVSCodeObject, scopeId: number = Math.random())=>{
