+++ src/client/src/utils/keyUpHandler.ts
@@ -1,5 +1,5 @@
 import * as React from "react";
-import {KEY_EVENTS} from "./constants"
+import {KEY_EVENTS} from "./constants/constants"
 
 const keyUpHandler = (e: React.KeyboardEvent<HTMLAnchorElement>) => {
